Adds patch-coverage to the build-process
This commit adds the ability to generate a patch-related
coverage-report. That way the code-coverage of the touched lines can be
checked. This would allow us to see what lines of the changed or added
code still are not tested without having to worry about all the old
code.

This currently only displays a message which lines aren't covered,
there's no failing build at the moment. I'm not sure whether that'd be a
good idea or not.

This is based on phploc (https://github.com/sebastianbergmann/phpcov)
which currently depends on PHPUnit 5.x. I haven't included this in the
package as I'm not sure on how to do this correctly and it would mean
updating phpunit.

YOu can install phploc f.e. via ```composer global require
phpunit/phpcov```